故障告警统计分析修改

+++ src/main/java/com/xiliulou/electricity/vo/TenantNoteRechargeVo.java
@@ -38,7 +38,7 @@ public class TenantNoteRechargeVo {
     /**
      * 短信充值次数
      */
-    private Integer rechargeNum;
+    private Long rechargeNum;
     
     /**
      * 充值时间
